+++ HW7/db.json
@@ -17,6 +17,12 @@
       "title": "Task 3",
       "isCompleted": true,
       "tag": "health"
+    },
+    {
+      "title": "Task 4",
+      "tag": "home",
+      "isCompleted": false,
+      "id": 4
     }
   ]
 }
