+++ HW7/src/components/header/Header.js
@@ -1,5 +1,5 @@
-import { useState } from '../../useState.js';
-import { useSearchString } from '../../useSearchString.js';
+import { useState } from '../../renders.js';
+import { useSearchString } from '../../renders.js';
 import { apiClient } from '../../clients/apiClient.js';
 import { SearchInput } from './SearchInput.js';
 import { Button } from '../Button.js'
