+++ HW7/src/components/header/TagForm.css
@@ -3,3 +3,48 @@
     width: 60%;
     margin-right: 130px;
 }
+.tag {
+    font-family: 'Inter';
+    padding: 2px 8px;
+    border-radius: 6px;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 15px;
+    margin-left: 20px;  
+    display: inline-block;
+}
+.healthTag {
+    background: rgba(60, 134, 244, 0.31);
+    color: #0053CF;
+}
+.homeTag {
+    background: #E2F7E2;
+    color: #639462;
+
+}
+.workTag {
+    background: #E8D7FF;
+    color: #9747FF;
+}
+.otherTag {
+    background: #FFECC7;
+    color: #EA8C00;
+}
+input[type="radio"]{
+    visibility: hidden;
+    height: 0;
+    width: 0;
+}
+input[type="radio"]:checked + label.homeTag{
+    border: 1px solid #639462;
+}
+input[type="radio"]:checked + label.healthTag{
+    border: 1px solid #0053CF;
+}
+input[type="radio"]:checked + label.workTag{
+    border: 1px solid #9747FF;
+}
+input[type="radio"]:checked + label.otherTag{
+    border: 1px solid #EA8C00;
+}
