+++ HW7/src/components/lists/App.js
@@ -1,5 +1,5 @@
-import { useSearchString } from '../../useSearchString.js';
-import { useState } from '../../useState.js';
+import { useSearchString } from '../../renders.js';
+import { useState } from '../../renders.js';
 import { apiClient } from '../../clients/apiClient.js';
 import { List } from './List.js';
 import { DoneList } from './DoneList.js';
