--- HW7/src/useSearchString.js
@@ -1,11 +1,0 @@
-let searchString = undefined;
-
-export function useSearchString(initialValue) {
-    searchString = searchString || initialValue;
-
-    function setValue(newValue) {
-        searchString = newValue;
-        renderApp();
-    }
-    return [searchString, setValue];
-}
