--- HW7/src/useState.js
@@ -1,11 +1,0 @@
-let state = undefined;
-
-export function useState(initialValue) {
-    state = state || initialValue;
-    function setValue(newValue) {
-        state = newValue;
-        renderApp();
-    
-    }
-    return [state, setValue];
-}
